Reinstate default Cypress screenshot behaviour
By default Cypress will capture a screenshot if a spec fails [1]. This is
useful when diagnosing spec failures especially in the CI build and we
have a step [2] in the CI GitHub Action that makes any such screenshots
available as downloadable build artefacts. Note that screenshots are
*not* captured by default when running `cypress open`, but that seems
sensible.

The default behaviour was changed in these two commits [3,4] so that
screenshots were no longer captured in the CI build. However, I can't
find any explanation beyond "Trying to fix screenshot problem".

I've reinstated the default behaviour in this commit.

[1]: https://docs.cypress.io/guides/guides/screenshots-and-videos#Screenshots
[2]: https://github.com/RaspberryPiFoundation/editor-ui/blob/2724a3c6db607eaa5e67bba9b730d9e1055ea467/.github/workflows/ci-cd.yml#L100-L106
[3]: https://github.com/RaspberryPiFoundation/editor-ui/commit/c209c3665ee6d17c31c47e314209983c90ff2921
[4]: https://github.com/RaspberryPiFoundation/editor-ui/commit/74290f5005441dc04e8ab8bd8611a5e8d7d3411a