Use quiet option when running Cypress specs in CI
This option [1] effectively enables the `--quiet` option [2] on the
`cypress run` CLI command which makes the output from the Cypress specs
a lot less verbose. This in turn makes it easier to see the wood for the
trees when a spec fails.

Note that it will still be possible to see verbose output when running
the specs locally.

[1]: https://github.com/cypress-io/github-action/blob/master/README.md#quiet-flag
[2]: https://docs.cypress.io/guides/guides/command-line#Options