Add e2e spec for remix behaviour used in editor-standalone
The editor-standalone app will:

* Listen for "editor-projectIdentifierChanged" custom events from the
  web component and will redirect to the remixed project's URL.
* Set the web component `load_remix_disabled` attribute to `true` so
  that the original project will be loaded in preference to the remixed
  version (unlike the way projects-ui works).

This spec checks that the web component provides the relevant behaviour
for editor-standalone to use.

I did attempt to add an example for the default behaviour (i.e. when
`load_remix_disabled` is not set). However, without any way to easily
reset the state of the projects in the editor-api database, it's hard to
come up with a robust way to write such a test.