Retain non-main components when setting code attribute
I think this was the original intention in the changes I've included
from #901, but there was a bug in the boolean logic.

The `otherComponents` should include all components other than `main.py`
& `index.html`. Previously it was typically not include any other
components, because the `component.extension !== defaultExtension`
expression was `false` for all of them, since in a Python project all
the files would have the `.py` (i.e. default) extension.