Bump version v0.23.0 -> v0.24.0 & update changelog
Strictly speaking, I think the fact we’ve removed a route [1] means we
should probably bump the major version, but given (a) we’re still on
v0.x.x which is less strict about semantic versioning; and (b) we’re
confident nothing is using that route any more, we've decided [2] to
just to bump the minor version, i.e. go for release v0.24.0.

[1]: https://github.com/RaspberryPiFoundation/editor-ui/pull/1013
[2]: https://raspberrypifoundation.slack.com/archives/C02JBAA2NFP/p1717429841916009