Add output_split_view attribute to web component
This is used to set the `isSplitView` state in the `EditorSlice` of the
Redux store which is in turn used in the `PythonRunner` to determine
whether to start by showing the split view or the tabbed view of the
output panel.

In the vast majority of cases in the editor app, including in the
embedded view, we want to start by displaying the split view. However,
@loiswells97 told me [1] that `isSplitView` was previously explicitly
set to `false` in the web component specifically for the Astro Pi
project [2] "to give the control panel and 3d model as much space as
possible". In some ways it would make more sense to have a
`output_tabbed_view` attribute defaulting to `false` and to change the
Astro Pi code to set this attribute to `true`. However, for now it seems
simpler to have this `output_split_view` attribute defaulting to `false`
so we don't have to change the Astro Pi code.

My actual motivation for making this change is to allow us to use the
web component in the embedded view in `editor-standalone` and obtain the
same behaviour as in with the embedded view in `editor-ui`. However, the
main `Project` & `MobileProject` components in `editor-standalone`
should also set `output_split_view` to `true` so that we obtain the same
behaviour for the main editor view as in the integrated editor in
`editor-ui`.

[1]: https://raspberrypifoundation.slack.com/archives/C02JBAA2NFP/p1717519701907459
[2]: https://missions.astro-pi.org/mz/code_submissions/new