Add boolean `embedded` attribute to web component
My motivation for adding this is to be able to use the web component in
the embedded view of editor-standalone. I'm planning on using this new
`embedded` attribute to indicate (among other things) whether a `RunBar`
component should be displayed *within* the `Output` component or not.

When we start using the web component in the embedded view in
editor-standalone, we'll want a `RunBar` to be displayed. However, in
the block-to-text app [1] we don't want a `RunBar` to be displayed. I've
already changed the latter so it doesn't try to set the `embedded`
attribute to `true` [2], so we should be OK to use `embedded` to make
the distinction.

Note that most of the infrastructure for this was already in place, I
just needed to add `embedded` to the list of boolean attributes in
`attributeChangedCallback`. For example, the value of the `embedded`
property in `WebComponentLoader` is already being used to set the
`isEmbedded` state via a call to `useEmbeddedMode()` [3].

[1]: https://github.com/RaspberryPiFoundation/block-to-text-alpha/
[2]: https://github.com/RaspberryPiFoundation/block-to-text-alpha/pull/125
[3]: https://github.com/RaspberryPiFoundation/editor-ui/blob/d9fd77eba4a9f5919c59cc82eec45df32fc5abe5/src/containers/WebComponentLoader.jsx#L83