Remove embedded & browserPreview props from Output
My motivation for making this change is to be able to use the web
component in the embedded view of editor-standalone. The `isEmbedded` &
`isBrowserPreview` states are used (among other things) to determine
whether a `RunBar` component should be displayed *within* the `Output`
component or not.

When we start using the web component in the embedded view in
editor-standalone, we'll want a `RunBar` to be displayed. I'm confident
that once I make some other changes in subsequent commit(s), it will be
sufficient to obtain the value of `isEmbedded` from the state and not
have to explicitly set it via a property on `Output`.

Note that this commit effectively reverts some changes that were made in
this commit [1].  I noted at the time that I didn't really understand
why we needed to make them.

[1]: https://github.com/RaspberryPiFoundation/editor-ui/commit/0d456b37a873d4fb1f37abf45ff2a4a59d4c32c8