Ignore webComponent state for RunBar in Output
My motivation for making this change is to be able to use the web
component in the embedded view of editor-standalone. When we are in this
embedded view, we will always want to display a `RunBar` component
within the `Output` component unless we are in browser preview mode
which is triggered by the "Preview in new tab" link in the output panel
of the main editor view.

Previously we were relying on the extra check on the `webComponent`
state to ensure the `RunBar` was *not* displayed in the block-to-text
app [1]. However, since I made this change [2], the `embedded` attribute
of the web component will no longer be set to `true` and so the
`isEmbedded` state will be `false` here. Thus the checks on the
`isEmbedded` & `isBrowserPreview` states will be sufficient to determine
whether to display the `RunBar` and we no longer need to check the
`webComponent` state.

[1]: https://github.com/RaspberryPiFoundation/block-to-text-alpha/
[2]: https://github.com/RaspberryPiFoundation/block-to-text-alpha/pull/125