Ignore webComponent state for preview link in HtmlRunner
My motivation for making this change is to be able to use the web
component in the embedded view of editor-standalone. When we are in this
embedded view, we will never want to display the "Preview in new tab"
link within the `HtmlRunner` component, i.e. within the `Output`
component.

Previously we were relying on the extra check on the `webComponent`
state to ensure the link was not displayed in the the block-to-text app
[1]. However, since I made this change [2], the `embedded` attribute of
the web component will no longer be set to `true` and so the
`isEmbedded` state will be `false` here. Thus the check on the
`isEmbedded` state will be sufficient to determine whether to display
the link and we no longer need to check the `webComponent` state.

[1]: https://github.com/RaspberryPiFoundation/block-to-text-alpha/
[2]: https://github.com/RaspberryPiFoundation/block-to-text-alpha/pull/125