WIP: Add scratch
NOTE that the Scratch GUI is rendering but with many errors and without
the assets that are currently missing.

Add ./scratch folder that attempts to generate the JS we need to embed
the Scratch editor in the iframe that renders scratch-component.html.

./scratch/src/index.jsx is a modified copy of both:

- https://github.com/scratchfoundation/scratch-gui/blob/141427d43db9e4471f8b628daa9f7c0c2c6d795f/src/playground/index.jsx
- https://github.com/scratchfoundation/scratch-gui/blob/141427d43db9e4471f8b628daa9f7c0c2c6d795f/src/playground/render-gui.jsx

Use `yarn start` to use webpacker dev server to make the editor
available at localhost:3011.

Use `yarn build` to generate the scratch/dist/main.js file that's then
copied by Webpack and required in scratch-component.html.