Update README.md to include how to initial install (#1248)
This pull request updates the project documentation to clarify the
recommended way to install dependencies.
It adds instructions to use Yarn 3 instead of npm due to compatibility
issues with some legacy packages.

Documentation improvements:

* Added a new section in `README.md` explaining that dependencies should
be installed using Yarn 3, and warning that `npm install` may fail
because of strict peer-dependency resolution.